cam_iface: fix Prosilica drivers
1. Return camera-based time estimate. Even if this has an unknown
offset and gain compared to the computer's clock, it should be
monotonic with actual camera frame number. Hence, it could be used to
infer camera frame sequence, which is not necessarily possible
otherwise (see below).

2. Add a semi-hack to attempt to calculate frame number even after
camera's estimate of frame number has rolled over back to 1. This is
complicated by the fact that the frames don't necessarily arrive in
order to the computer.


git-svn-id: https://code.astraw.com/motmot/trunk/cam_iface@716 9608957c-362e-0410-90bb-ddb82fccb57f