Fix many (all) enumeration / race bugs
The symptom was corrupted communications when enumerating
many cameras.

The root cause was during enumeration there is multiple
arv_cameras created and never unreffed and importantly,
the stream is never started on them. While I am not
convinced this isnt a bug in aravis (unstarted cameras should
not eat data), it is safest to unref and thus cleanup those
cameras we enumerate but do not start streaming with.

Unfortunately this means we lose some cam_info granualarity,
(I don't fill out the vendor and model, only the chip/guid),
but at least it works now