API/ABI Break
Add an additional construction parameter - a const char *
to specify the interface on the host machine used to communicate
with the camera. The meaning is backend dependent. For example,
in the aravis case, the interface is a ip address of the interface
used to communicate with the camera.

It is OK to pass NULL for this value.

Only the aravis backend understands this new argument.